Length passed to nhash is 32-bit only.y
git-svn-id: svn://svn.code.sf.net/p/wsjt/wsjt/branches/wsjtx_exp@5794 ab8295b8-cf94-4d9e-aec4-7959e3be5d79

+++ lib/nhash.c
@@ -374,7 +374,8 @@ uint32_t nhash( const void *key, size_t length, uint32_t initval)
 /*
  * Fortran argument compatible wrapper
  */
-uint32_t nhash_( const void * key, size_t const * length, uint32_t const * initval)
+//uint32_t nhash_( const void * key, size_t const * length, uint32_t const * initval)
+uint32_t nhash_( const void * key, uint32_t const * length, uint32_t const * initval)
 {
   return nhash (key, *length, *initval);
 }
